MetricsAgentTests/AgentControllerUnitTest.cs
using MetricsAgent.Controllers;
using Microsoft.AspNetCore.Mvc;
using System;
using MetricsManager.Enums;
using Xunit;

namespace MetricsAgentTests
{
    public class AgentControllerUnitTest
    {
        private readonly AgentController _controller;

        public AgentControllerUnitTest()
        {
            _controller = new AgentController();
        }
        
        [Fact]
        public void GetCpuMetricsByPercentile_ReturnsOk()
        {
            //Arrange
            var fromTime = TimeSpan.FromSeconds(0);
            var toTime = TimeSpan.FromSeconds(100);
            var percentile = Percentile.P95;

            //Act
            var result = _controller.GetCpuMetricsByPercentile(fromTime, toTime, percentile);

            // Assert
            _ = Assert.IsAssignableFrom<IActionResult>(result);
        }
        
        [Fact]
        public void GetCpuMetrics_ReturnsOk()
        {
            //Arrange
            var fromTime = TimeSpan.FromSeconds(0);
            var toTime = TimeSpan.FromSeconds(100);

            //Act
            var result = _controller.GetCpuMetrics(fromTime, toTime);

            // Assert
            _ = Assert.IsAssignableFrom<IActionResult>(result);
        }
        
        [Fact]
        public void GetDotNetErrorsCount_ReturnsOk()
        {
            //Arrange
            var fromTime = TimeSpan.FromSeconds(0);
            var toTime = TimeSpan.FromSeconds(100);

            //Act
            var result = _controller.GetDotNetErrorsCount(fromTime, toTime);

            // Assert
            _ = Assert.IsAssignableFrom<IActionResult>(result);
        }
        
        [Fact]
        public void GetNetworkMetrics_ReturnsOk()
        {
            //Arrange
            var fromTime = TimeSpan.FromSeconds(0);
            var toTime = TimeSpan.FromSeconds(100);

            //Act
            var result = _controller.GetNetworkMetrics(fromTime, toTime);

            // Assert
            _ = Assert.IsAssignableFrom<IActionResult>(result);
        }
        
        [Fact]
        public void GetHddLeftSpace_ReturnsOk()
        {
            //Arrange

            //Act
            var result = _controller.GetHddLeftSpace();

            // Assert
            _ = Assert.IsAssignableFrom<IActionResult>(result);
        }
        
        [Fact]
        public void GetRamAvailable_ReturnsOk()
        {
            //Arrange

            //Act
            var result = _controller.GetRamAvailable();

            // Assert
            _ = Assert.IsAssignableFrom<IActionResult>(result);
        }
    }
}